Fixed siret type : int -> bigint

+++ backend/database/schema.sql
@@ -19,7 +19,7 @@ create table business_sectors (
 create table enterprise (
   id varchar(36) not null,
   name varchar(100) not null,
-  siret int not null,
+  siret bigint not null,
   description text not null,
   email varchar(320) not null,
   password varchar(100) not null,
